移設と互換ラッパーの配置
